Split manager controllers away from monolithic app context
